Update modHelper.py
updated with feature 1 established: the bot replies with tele links adapted from https://telenus.nusmods.com/ once bot is called in the comment section of top 20 posts sorted by new. The program is currently set to run every hour, yet the actual running frequency can be changed according to needs. 

Note: r/pythonforengineers is a subreddit we are using for testing purposes as we do not want to spam and disturb r/NUS.