telebot 1.0 - a rough sketch of basic structure
This is a telegram bot aiming to fulfil features 2 and 3 - to match students and facilitate chats based on their needs and preferences. To protect privacy and ensure safety, matching and anonymous chat will be conducted through the bot and students may choose to reveal their usernames upon satisfactory matches. The current version sets up the basics of a telebot and extracts keywords about module and faculty based on students' inputs.